requirements.txt
langchain == 0.3.0
openai == 1.95.1
azure-search-documents == 11.5.3
pypdf == 4.3.1
streamlit == 1.46.1
python-dotenv == 1.0.1
azure-core == 1.34.0